+++ lesson23/src/main/java/model/User.java
@@ -1,3 +1,5 @@
+package model;
+
 public class User {
     private int id;
     private String username;
